Column changed to float to avoid anndata saving issue

+++ neuromast3d/visualization/analysis.py
@@ -193,7 +193,7 @@ def main():
 
         # If not from an experiment with that label, remain NaN
         binarized[adata.obs[f'{col_name}_z_score'].isnull()] = np.NaN
-        adata.obs[f'{ch_name}_positive'] = binarized
+        adata.obs[f'{ch_name}_positive'] = binarized.astype(float)
         sc.pl.umap(adata, color=f'{ch_name}_positive', save=f'{ch_name}_umap.png', show=False)
         sc.pl.umap(adata, color=f'{col_name}_z_score', save=f'{ch_name}_z_score.png', show=False)
